add test for smartengine map
