+++ crates/fluvio-smartengine/src/transforms/test.rs
@@ -10,14 +10,12 @@ use fluvio_smartmodule::{
 
 use crate::{SmartEngine, SmartModuleConfig, SmartModuleInitialData};
 
-const SM_FILTER: &str = "fluvio_smartmodule_filter";
-const SM_FILTER_INIT: &str = "fluvio_smartmodule_filter_init";
 const SM_ARRAY_MAP: &str = "fluvio_wasm_array_map_array";
 const SM_FILTER_MAP: &str = "fluvio_wasm_filter_map";
 const SM_AGGEGRATE: &str = "fluvio_wasm_aggregate";
 const SM_JOIN: &str = "fluvio_wasm_join";
 
-fn read_wasm_module(module_name: &str) -> Vec<u8> {
+pub(crate) fn read_wasm_module(module_name: &str) -> Vec<u8> {
     let spu_dir = std::env::var("CARGO_MANIFEST_DIR").expect("target");
     let wasm_path = PathBuf::from(spu_dir)
         .parent()
@@ -29,151 +27,9 @@ fn read_wasm_module(module_name: &str) -> Vec<u8> {
     read_module_from_path(wasm_path)
 }
 
-fn read_module_from_path(filter_path: impl AsRef<Path>) -> Vec<u8> {
+pub(crate) fn read_module_from_path(filter_path: impl AsRef<Path>) -> Vec<u8> {
     let path = filter_path.as_ref();
     std::fs::read(path).unwrap_or_else(|_| panic!("Unable to read file {}", path.display()))
-}
-
-#[ignore]
-#[test]
-fn test_filter() {
-    let engine = SmartEngine::new();
-    let mut chain_builder = engine.builder();
-
-    chain_builder
-        .add_smart_module(
-            SmartModuleConfig::builder().build().unwrap(),
-            read_wasm_module(SM_FILTER),
-        )
-        .expect("failed to create filter");
-
-    assert_eq!(
-        chain_builder
-            .instances()
-            .first()
-            .expect("first")
-            .transform()
-            .name(),
-        crate::transforms::filter::FILTER_FN_NAME
-    );
-
-    let mut chain = chain_builder.initialize().expect("failed to build chain");
-
-    let input = vec![Record::new("hello world")];
-    let output = chain
-        .process(SmartModuleInput::try_from(input).expect("input"))
-        .expect("process");
-    assert_eq!(output.successes.len(), 0); // no records passed
-
-    let input = vec![Record::new("apple"), Record::new("fruit")];
-    let output = chain
-        .process(SmartModuleInput::try_from(input).expect("input"))
-        .expect("process");
-    assert_eq!(output.successes.len(), 1); // one record passed
-    assert_eq!(output.successes[0].value.as_ref(), b"apple");
-}
-
-#[ignore]
-#[test]
-fn test_filter_with_init_invalid_param() {
-    let engine = SmartEngine::new();
-    let mut chain_builder = engine.builder();
-
-    chain_builder
-        .add_smart_module(
-            SmartModuleConfig::builder().build().unwrap(),
-            read_wasm_module(SM_FILTER_INIT),
-        )
-        .expect("failed to create filter");
-
-    assert_eq!(
-        chain_builder
-            .instances()
-            .first()
-            .expect("first")
-            .transform()
-            .name(),
-        crate::transforms::filter::FILTER_FN_NAME
-    );
-
-    assert_eq!(
-        chain_builder
-            .initialize()
-            .expect_err("should return param error")
-            .to_string(),
-        "Missing param key\n\nSmartModule Init Error: \n"
-    );
-}
-
-#[ignore]
-#[test]
-fn test_filter_with_init_ok() {
-    let engine = SmartEngine::new();
-    let mut chain_builder = engine.builder();
-
-    chain_builder
-        .add_smart_module(
-            SmartModuleConfig::builder()
-                .param("key", "a")
-                .build()
-                .unwrap(),
-            read_wasm_module(SM_FILTER_INIT),
-        )
-        .expect("failed to create filter");
-
-    let instance = chain_builder.instances().first().expect("first");
-
-    assert_eq!(
-        instance.transform().name(),
-        crate::transforms::filter::FILTER_FN_NAME
-    );
-
-    assert!(instance.get_init().is_some());
-
-    let mut chain = chain_builder.initialize().expect("failed to build chain");
-
-    let input = vec![Record::new("hello world")];
-    let output = chain
-        .process(SmartModuleInput::try_from(input).expect("input"))
-        .expect("process");
-    assert_eq!(output.successes.len(), 0); // no records passed
-
-    let input = vec![
-        Record::new("apple"),
-        Record::new("fruit"),
-        Record::new("banana"),
-    ];
-    let output = chain
-        .process(SmartModuleInput::try_from(input).expect("input"))
-        .expect("process");
-    assert_eq!(output.successes.len(), 2); // one record passed
-    assert_eq!(output.successes[0].value.as_ref(), b"apple");
-    assert_eq!(output.successes[1].value.as_ref(), b"banana");
-
-    // build 2nd chain with different parameter
-    let mut chain_builder = engine.builder();
-    chain_builder
-        .add_smart_module(
-            SmartModuleConfig::builder()
-                .param("key", "b")
-                .build()
-                .unwrap(),
-            read_wasm_module(SM_FILTER_INIT),
-        )
-        .expect("failed to create filter");
-
-    let mut chain = chain_builder.initialize().expect("failed to build chain");
-
-    let input = vec![
-        Record::new("apple"),
-        Record::new("fruit"),
-        Record::new("banana"),
-    ];
-    let output = chain
-        .process(SmartModuleInput::try_from(input).expect("input"))
-        .expect("process");
-    assert_eq!(output.successes.len(), 1); // only banana
-    assert_eq!(output.successes[0].value.as_ref(), b"banana");
 }
 
 #[ignore]
