+++ crates/fluvio-smartmodule/examples/map/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "fluvio-wasm-map"
+name = "fluvio-smartmodule-map"
 version = "0.0.0"
 authors = ["Fluvio Contributors <team@fluvio.io>"]
 edition = "2021"
